# Slight performance improvement for `IEdmNavigationSource.EntityType()`