--- a/src/Microsoft.OData.Edm/ExtensionMethods/ExtensionMethods.cs
+++ b/src/Microsoft.OData.Edm/ExtensionMethods/ExtensionMethods.cs
@@ -2822,34 +2822,7 @@ namespace Microsoft.OData.Edm
         /// <returns>The entity type of the navigation source.</returns>
         public static IEdmEntityType EntityType(this IEdmNavigationSource navigationSource)
         {
-            var entitySetBase = navigationSource as IEdmEntitySetBase;
-            if (entitySetBase != null)
-            {
-                IEdmCollectionType collectionType = entitySetBase.Type as IEdmCollectionType;
-
-                if (collectionType != null)
-                {
-                    return collectionType.ElementType.Definition as IEdmEntityType;
-                }
-
-                var unknownEntitySet = entitySetBase as IEdmUnknownEntitySet;
-                if (unknownEntitySet != null)
-                {
-                    // Handle missing navigation target for nullable
-                    // singleton navigation property.
-                    return unknownEntitySet.Type as IEdmEntityType;
-                }
-
-                return null;
-            }
-
-            var singleton = navigationSource as IEdmSingleton;
-            if (singleton != null)
-            {
-                return singleton.Type as IEdmEntityType;
-            }
-
-            return null;
+            return navigationSource?.Type.AsElementType() as IEdmEntityType;
         }
 
         #endregion